perf: 총 당첨 상금 계산 로직을 수정하라
총 당첨 상금을 double 타입으로 계산하지 않고 long 타입으로 계산한다.

Closes

+++ src/main/java/lotto/service/LottoDraw.java
@@ -43,15 +43,15 @@ public class LottoDraw {
         }
     }
 
-    public double calcTotalPrize() {
-        double totalPrize = 0.0;
+    public long calcTotalPrize() {
+        long totalPrize = 0L;
         for (LottoRank key : LottoRank.values()) {
             totalPrize += LottoWinningRule.getPrize(key) * statistics.search(key);
         }
         return totalPrize;
     }
 
-    public double calcReturnRate(double totalPrize, Amount amount) {
-        return Math.round(totalPrize / amount.getValue() * PERCENTAGE * DECIMAL_PLACE) / DECIMAL_PLACE;
+    public double calcReturnRate(long totalPrize, Amount amount) {
+        return Math.round((double) totalPrize / amount.getValue() * PERCENTAGE * DECIMAL_PLACE) / DECIMAL_PLACE;
     }
 }
